order code는 완료. test로 수정 필요

+++ src/main/java/com/example/plantmall/domain/LineItem.java
@@ -24,7 +24,7 @@ public class LineItem implements Serializable {
 	    this.productId = cartItem.getProduct().getProductId();
 	}
 	
-	public double getTotalPrice() {
+	public int getTotalPrice() {
 		return this.unitPrice * this.quantity;
 	}
 }
